+++ packages/expo-file-system/android/src/main/java/expo/modules/filesystem/FileSystemModule.kt
@@ -14,6 +14,7 @@ import android.util.Base64
 import android.util.Log
 import androidx.core.content.FileProvider
 import androidx.documentfile.provider.DocumentFile
+import expo.modules.core.errors.InvalidArgumentException
 import expo.modules.core.errors.ModuleDestroyedException
 import expo.modules.interfaces.filesystem.Permission
 import expo.modules.kotlin.Promise
@@ -823,23 +824,34 @@ open class FileSystemModule : Module() {
     if (!documentFile.exists()) {
       return
     }
-    if (!outputDir.exists() && !outputDir.mkdirs()) {
+    if (!outputDir.isDirectory) {
+      outputDir.parentFile?.let {
+        if (!it.exists() && !it.mkdirs()) {
+          throw IOException("Couldn't create folder in output dir.")
+        }
+      }
+    } else if (!outputDir.exists() && !outputDir.mkdirs()) {
       throw IOException("Couldn't create folder in output dir.")
     }
+
     if (documentFile.isDirectory) {
       for (file in documentFile.listFiles()) {
-        documentFile.name?.let {
-          transformFilesFromSAF(file, File(outputDir, it), copy)
-        }
+        transformFilesFromSAF(file, outputDir, copy)
       }
       if (!copy) {
         documentFile.delete()
       }
       return
     }
+
     documentFile.name?.let {
-      val newFile = File(outputDir.path, it)
-      context.contentResolver.openInputStream(documentFile.uri).use { `in` -> FileOutputStream(newFile).use { out -> IOUtils.copy(`in`, out) } }
+      val newFile = if (outputDir.isDirectory) {
+        File(outputDir.path, it)
+      } else {
+        File(outputDir.path)
+      }
+      context.contentResolver.openInputStream(documentFile.uri)
+        .use { `in` -> FileOutputStream(newFile).use { out -> IOUtils.copy(`in`, out) } }
       if (!copy) {
         documentFile.delete()
       }